Update Jenkinsfile2
SonarQube for java needs the class files or code coverage analysis. In a "full" pipeline that is build and tested, the class files are stashed in the build process and un stashed in the latter ones.

They can also be gathered from Nexus, but it's better to follow the aproach of the sonarqube maven plugin and just build again.